Refactor Directory * Move all class files into `/classes` module
