+++ source_collectors/muckrock/classes/FOIASearcher.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from source_collectors.muckrock.muckrock_fetchers.FOIAFetcher import FOIAFetcher
+from source_collectors.muckrock.classes.muckrock_fetchers import FOIAFetcher
 from tqdm import tqdm
 
 class FOIASearcher:
